requirements.txt
# 项目依赖列表

python-docx>=0.8.11
python-dotenv>=0.19.0
openai>=1.0.0
anthropic>=0.5.0
markdown>=3.4.0
pillow>=8.3.0
requests>=2.28.0
weasyprint>=51.0
flask>=2.0.0
psutil>=5.9.0